Validate comment site URL
Since people will share jekyll site templates, it'd be very easy for someone to accidentally host a site that posts to someone else's Jekyll receiver. We want to avoid that happening by mistake. It's still possible for someone to post to another receiver, but they'll have to be intentional about it, and it won't really do them any good.

This requires that comment forms add the following field to their template: <input name="comment-site" type="hidden" value="{{ site.url }}">